+++ packages/common/tests/DateUtil.test.ts
@@ -3,6 +3,8 @@ import { DateUtil } from '../src/utils/DateUtil.js'
 
 const fakeDateTime = new Date(2023, 1, 1, 12)
 const fakeDateTimeOneHourAgo = new Date(2023, 1, 1, 11)
+const fakeDateTimeOneTwoMinsAgo = new Date(2023, 1, 1, 12, 2)
+const fakeDateTimeOneSecondAgo = new Date(2023, 1, 1, 12, 0, 30)
 
 // -- Tests --------------------------------------------------------------------
 describe('DateUtil', () => {
@@ -24,6 +26,18 @@ describe('DateUtil', () => {
     vi.setSystemTime(fakeDateTime)
 
     expect(DateUtil.getRelativeDateFromNow(fakeDateTimeOneHourAgo.getTime())).toEqual('1 hr')
+  })
+
+  it('should return relative time for minutes ago', () => {
+    vi.setSystemTime(fakeDateTime)
+
+    expect(DateUtil.getRelativeDateFromNow(fakeDateTimeOneTwoMinsAgo.getTime())).toEqual('2 min')
+  })
+
+  it('should return relative time for seconds ago', () => {
+    vi.setSystemTime(fakeDateTime)
+
+    expect(DateUtil.getRelativeDateFromNow(fakeDateTimeOneSecondAgo.getTime())).toEqual('30 sec')
   })
 
   it('should format date correctly', () => {