+++ packages/core/src/controllers/AccountController.ts
@@ -5,6 +5,10 @@ import type { AccountType, CaipAddress, ConnectedWalletInfo } from '../utils/Typ
 import type { Balance } from '@web3modal/common'
 import { BlockchainApiController } from './BlockchainApiController.js'
 import { SnackController } from './SnackController.js'
+import { SwapController } from './SwapController.js'
+import { SwapApiUtil } from '../utils/SwapApiUtil.js'
+import type { W3mFrameTypes } from '@web3modal/wallet'
+import { NetworkController } from './NetworkController.js'
 
 // -- Types --------------------------------------------- //
 export interface AccountControllerState {
@@ -23,6 +27,7 @@ export interface AccountControllerState {
   tokenBalance?: Balance[]
   shouldUpdateToAddress?: string
   connectedWalletInfo?: ConnectedWalletInfo
+  preferredAccountType?: W3mFrameTypes.AccountType
 }
 
 type StateKey = keyof AccountControllerState
@@ -107,12 +112,19 @@ export const AccountController = {
     state.connectedWalletInfo = connectedWalletInfo
   },
 
+  setPreferredAccountType(preferredAccountType: AccountControllerState['preferredAccountType']) {
+    state.preferredAccountType = preferredAccountType
+  },
+
   async fetchTokenBalance() {
+    const chainId = NetworkController.state.caipNetwork?.id
+
     try {
-      if (state.address) {
-        const response = await BlockchainApiController.getBalance(state.address)
+      if (state.address && chainId) {
+        const response = await BlockchainApiController.getBalance(state.address, chainId)
 
         this.setTokenBalance(response.balances)
+        SwapController.setBalances(SwapApiUtil.mapBalancesToSwapTokens(response.balances))
       }
     } catch (error) {
       SnackController.showError('Failed to fetch token balance')
@@ -133,5 +145,6 @@ export const AccountController = {
     state.tokenBalance = []
     state.shouldUpdateToAddress = undefined
     state.connectedWalletInfo = undefined
+    state.preferredAccountType = undefined
   }
 }