+++ packages/scaffold/src/partials/w3m-tooltip-trigger/styles.ts
@@ -1,0 +1,8 @@
+import { css } from 'lit'
+
+export default css`
+  :host {
+    width: 100%;
+    display: block;
+  }
+`