+++ packages/solana/src/connectors/walletAdapters.ts
@@ -1,7 +1,9 @@
-import { PhantomWalletAdapter } from '@solana/wallet-adapter-phantom'
-import { SolflareWalletAdapter } from '@solana/wallet-adapter-solflare'
-import { TrustWalletAdapter } from '@solana/wallet-adapter-trust'
-import { BackpackWalletAdapter } from '@solana/wallet-adapter-backpack'
+import {
+  PhantomWalletAdapter,
+  BackpackWalletAdapter,
+  SolflareWalletAdapter,
+  TrustWalletAdapter
+} from './walletAdapters/index.js'
 
 import type { BaseWalletAdapter } from '@solana/wallet-adapter-base'
 import type { Connector } from '@web3modal/scaffold'