+++ packages/solana/src/connectors/walletAdapters/TrustWalletAdapter.ts
@@ -1,0 +1,6 @@
+import { TrustWalletAdapter as SolanaTrustWalletAdapter } from '@solana/wallet-adapter-trust'
+import { signAndSendTransaction } from './utils.js'
+
+export class TrustWalletAdapter extends SolanaTrustWalletAdapter {
+  signAndSendTransaction = signAndSendTransaction(this)
+}