+++ packages/ethers5/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @web3modal/ethers5
 
+## 4.1.12-910a844.0
+
+### Patch Changes
+
+- refactor: sync theme with secure site
+
+- Updated dependencies []:
+  - @web3modal/scaffold@4.1.12-910a844.0
+  - @web3modal/polyfills@4.1.12-910a844.0
+  - @web3modal/scaffold-react@4.1.12-910a844.0
+  - @web3modal/scaffold-utils@4.1.12-910a844.0
+  - @web3modal/scaffold-vue@4.1.12-910a844.0
+  - @web3modal/siwe@4.1.12-910a844.0
+
+## 4.1.11
+
+### Patch Changes
+
+- refactor: Solana exported helper methods and proram instruction improvements
+- refactor: email and smart account improvements
+- refactor: unit test CI check improvements
+- feat: convert
+
+- Updated dependencies []:
+  - @web3modal/polyfills@4.1.11
+  - @web3modal/scaffold@4.1.11
+  - @web3modal/scaffold-react@4.1.11
+  - @web3modal/scaffold-utils@4.1.11
+  - @web3modal/scaffold-vue@4.1.11
+  - @web3modal/siwe@4.1.11
+
 ## 4.1.10
 
 ### Patch Changes
@@ -69,6 +100,18 @@
   - @web3modal/scaffold-utils@4.1.6
   - @web3modal/scaffold-vue@4.1.6
   - @web3modal/siwe@4.1.6
+
+## 4.1.6-a0733f5.0
+
+### Patch Changes
+
+- chore: canary release for siwe dependency
+
+- Updated dependencies []:
+  - @web3modal/ethers@4.1.6-a0733f5.0
+  - @web3modal/siwe@4.1.6-a0733f5.0
+  - @web3modal/solana@4.1.6-a0733f5.0
+  - @web3modal/wagmi@4.1.6-a0733f5.0
 
 ## 4.1.5